refacto: Adding Modules in structure
