--- src/MyFITJob.Api/MyFITJob.Api/JobOffers/Domain/DisplayNameAttribute.cs
@@ -1,12 +1,0 @@
-namespace MyFITJob.Api.JobOffers.Domain;
-
-[AttributeUsage(AttributeTargets.Field)]
-public class DisplayNameAttribute : Attribute
-{
-    public string DisplayName { get; }
-
-    public DisplayNameAttribute(string displayName)
-    {
-        DisplayName = displayName;
-    }
-}
