+++ src/MyFITJob.Api/MyFITJob.Api/Modules/JobOffers/Domain/JobOffer.cs
